+++ cmd/racedetector/main.go
@@ -113,18 +113,4 @@ FOR MORE INFORMATION:
 
 // buildCommand is implemented in build.go
 // runCommand is implemented in run.go
-
-// testCommand implements the 'racedetector test' command.
-//
-// This will instrument and run tests with race detection enabled.
-// Currently not implemented - planned for v0.4.0.
-//
-// Example:
-//
-//	racedetector test -v ./internal/...
-func testCommand(_ []string) {
-	// TODO: Implement test command in v0.4.0 (args unused until implemented)
-	fmt.Fprintln(os.Stderr, "Error: 'test' command not yet implemented")
-	fmt.Fprintln(os.Stderr, "This will be implemented in v0.4.0")
-	os.Exit(1)
-}
+// testCommand is implemented in test.go